+++ packages/ubcli/lib/createStore.js
@@ -63,7 +63,6 @@ module.exports = function createStore (options) {
   }
 
   function createOneStore (cStore) {
-    console.log('Start handle blobStore "%s"', cStore.name)
     if (!cStore.storeType) {
       cStore.storeType = 'FileSystem'
     }
@@ -72,9 +71,8 @@ module.exports = function createStore (options) {
       if (!RE_TRAILING_PATH_SEP.test(cStorePath)) {
         cStorePath += path.sep
       }
-      console.log('\tresolved to path', cStorePath)
       if (!fs.existsSync(cStorePath)) {
-        console.log('\tresolved path not exists. Do force directory')
+        console.log(`\tStore ${cStore.name}: path '${cStorePath}' not exists and will be created`)
         fs.mkdirSync(cStorePath)
       }
     } else {
@@ -83,11 +81,11 @@ module.exports = function createStore (options) {
     if (cStore.tempPath) {
       const tmp = cStore.tempPath // already converted to absolute by argv
       if (!fs.existsSync(tmp)) {
-        console.log('\t Create temp directory %s', tmp)
+        console.log(`\tStore ${cStore.name}: create temp directory ${tmp}`)
         fs.mkdirSync(tmp)
       }
     }
-    console.log('Done!')
+    console.log('BLOB stores path is OK')
   }
 
   selectedStores.forEach(createOneStore)